requirements.txt
telethon==1.34.0
python-dotenv==1.0.0
aiohttp==3.9.1
aiofiles==23.2.1
asyncio-mqtt==0.16.1
